Move the GC metrics out to their own file
Move the manual GC code into its own function, and then move that, and all the
other GC stuff, out to a separate file.
